clean fluid.layers.controw_flow
